==> Scripts/Analysis/3-models_health-military.qmd <==
---
title : "Analysis -  Military health"
editor: visual
date  : "March 8, 2023"
---

## Purpose

Estimate military health aid models.

## Setup

```{r}
#| warning: false
#| message: false
# ----------------------------------- #
# Clear working environment  
# ----------------------------------- #
rm(list = ls())
# ----------------------------------- #

# ----------------------------------- #
# Load required packages
# ----------------------------------- #
library(tidyverse)
library(sf)
library(texreg)
library(sandwich)
library(lmtest)
library(margins)
library(kableExtra)
# ----------------------------------- #

# ----------------------------------- #
# Load data
# ----------------------------------- #
d <- read_rds("../../Data/analysis-data/0-analysis_data.rds")
# ----------------------------------- #
```

## Military health

### Formula

```{r}
base    <-  . ~ poly1 + poly2 + poly3 + region 
full    <-  update(base, ~ . + un_gdppc + un_gdppc2 + 
                     un_lifeexpec_yrs + un_lifeexpec_yrs2 + un_infdeath_p1000 + 
                     ghdx_ncd_prevalence_pc + ghdx_cmd_prevalence_pc + 
                     emdat_disasters_peryr + ged_major + atop_nonagg + 
                     trade_net + vdem + vdem2)

fs <- list(base = base, full = full)
rm(base, full)
```

### Fit

```{r}
us <- lapply(fs, function(x){
  glm(formula = update(x, mil_health_us ~ mil_health_cn + .),
      data    = d, 
      family  = binomial(link = "probit"))
})

cn <- lapply(fs, function(x){
  glm(formula = update(x, mil_health_cn ~ mil_health_us + .),
      data    = d, 
      family  = binomial(link = "probit"))
})

# Models list
mods_military <- list("us" = us, "cn" = cn) %>% list_flatten()

# Goodness of fit statistics
mods_military_gof <- list(
  "FE Region"      = rep("Yes", length(mods_military)),
  "Time polynomial"= rep("Yes", length(mods_military)),
  "Obs."           = lapply(mods_military, function(x) nrow(model.matrix(x))),
  "Log Likelihood" = lapply(mods_military, function(x) round(logLik(x) %>% as.numeric(), 2))
)

rm(us, cn)
```

### Robust SEs

```{r}
mods_military_robust <- lapply(mods_military, vcovPC, cluster = ~ ccode)
```

## Save

```{r}
save(mods_military, mods_military_gof, mods_military_robust, 
     file = "../../Data/analysis-data/3-models_health-military.Rdata")

# rm(list = ls())
```

------------------------------------------------------------------------

# \_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_\_

## Robustness checks

-   Additional models to explore robustness of results

Note - this subsetting block tests robustness of results by retaining only those countries in the data which have at least 1 year of either US or Chinese military health aid. Main results retain their significance at the 0.1 level.

### Subset - at least 1 year mil aid

```{r}
dsub <- d %>%
  st_drop_geometry() %>%
  group_by(cname) %>%
  summarize(tgt = ifelse(any(mil_health_cn == 1) | any(mil_health_us == 1), 1, 0)) %>%
  ungroup %>%
  filter(tgt == 1) %>%
  pull(cname)

dsub <- d %>%
  filter(cname %in% dsub)
```

### SAR Probits

```{r}
#| eval: false

library(spatialprobit)

load("Data/analysis-data/0-spatial_weights.rdata")
w <- Matrix(d_spatial$w_knn, sparse = TRUE)

?sarprobit
us <- sarprobit(formula = update(fs$full, mil_health_us ~ mil_health_us_lag + 
                                   mil_health_cn + . - region),
                W       = w,
                data    = d,
                ndraw   = 2e3,
                burn.in = 1e3)

summary(us)

cn <- sarprobit(formula = update(fs$full, mil_health_cn ~ mil_health_cn_lag + 
                                   mil_health_us + . - region),
                W       = w,
                data    = d,
                ndraw   = 2e3,
                burn.in = 1e3)

summary(cn)
```

Concerning military model fit:

-   With contemporaneous rival -- US and China significant
-   With contemporaneous rival + lag rival - Only US significant
-   With contemporaneous rival + lag rival + own lag - Only US significant
-   With contemporaneous rival + own lag - **Both significant**
    -   NOTE - not robust to inclusion of year FE - but probably overkill with time poly.

    -   Plus - based on error appears separation occurs with year FE.

### Multinomial analysis

Analysis using `mil_health_qual` variable.

```{r}
# library(MNP)
# d <- read_rds("../../Data/analysis-data/0-analysis_data.rds") %>%
#   st_drop_geometry %>%
#   group_by(ccode) %>%
#   mutate(across(.cols = c(mil_health_us, mil_health_cn),
#                 .fns  = ~ .x - lag(.x, n = 1),
#                 .names= "{.col}_fd")) %>%
#   ungroup %>%
#   drop_na(mil_health_cn_fd, mil_health_us_fd) %>%
#   mutate(across(.cols = c(mil_health_cn_fd, mil_health_us_fd),
#                 .fns  = ~factor(.x, levels = c(0, -1, 1), labels = c("no_change", "lean_out", "lean_in")))) %>% 
#   select(ccode, year, starts_with("mil_health"), 
#          un_gdppc, un_gdppc2, vdem, vdem2, region)
# 
# 
# full    <-  . ~ un_gdppc + un_gdppc2 + vdem + vdem2 + region
# 
# m <- mnp(formula = update(full, mil_health_cn_fd ~ mil_health_us_lag + .),
#          data    = d, n.draws = 2e3, burnin = 1e3)
# m <- mnp(formula = update(full, mil_health_us_fd ~ mil_health_cn_lag + .),
#          data    = d, n.draws = 2e3, burnin = 1e3)
# summary(m)

# alternative
# mm <- glm(mil_health_us_fd == "lean_in" ~ mil_health_cn + mil_health_cn_lag + un_gdppc + un_gdppc2 + vdem + vdem2 + region + year, data = d, family = binomial(link = "probit"))
```

### Bivariate probit analysis

-   with `mvProbit` package

```{r}
# library(mvProbit)
# m <- mvProbit(formula = cbind(mil_health_us, mil_health_cn) ~ un_gdppc + un_gdppc2 + un_infdeath_p1000, data = d)
```

-   with `VGAM` package

```{r}
# library(VGAM)
# mm <- vglm(formula = cbind(mil_health_us, mil_health_cn) ~ un_gdppc + un_gdppc2 + 
#              un_lifeexpec_yrs + un_lifeexpec_yrs2 +
#              un_infdeath_p1000 + poly1 + poly2 + poly3, 
#            data    = d, 
#            family  = binom2.or)
```
